+++ compiler-plugin/src/main/java/io/ballerina/stdlib/graphql/compiler/validator/errors/ErrorCode.java
@@ -40,5 +40,8 @@ public enum ErrorCode {
     GRAPHQL_116,
     GRAPHQL_117,
     GRAPHQL_118,
-    GRAPHQL_119
+    GRAPHQL_119,
+    GRAPHQL_120,
+    GRAPHQL_121,
+    GRAPHQL_122
 }